# zip-repl/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymFlow Gym Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemberMembershipType =
  (typeof MemberMembershipType)[keyof typeof MemberMembershipType];

export const MemberMembershipType = {
  basic: "basic",
  standard: "standard",
  premium: "premium",
} as const;

export type MemberStatus = (typeof MemberStatus)[keyof typeof MemberStatus];

export const MemberStatus = {
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export interface Member {
  id: number;
  name: string;
  email: string;
  phone: string;
  membershipType: MemberMembershipType;
  membershipExpiry: string;
  status: MemberStatus;
  joinedAt: string;
  /** @nullable */
  workoutPlanId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type MemberInputMembershipType =
  (typeof MemberInputMembershipType)[keyof typeof MemberInputMembershipType];

export const MemberInputMembershipType = {
  basic: "basic",
  standard: "standard",
  premium: "premium",
} as const;

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  membershipType: MemberInputMembershipType;
  membershipExpiry: string;
  /** @nullable */
  workoutPlanId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type MemberUpdateMembershipType =
  (typeof MemberUpdateMembershipType)[keyof typeof MemberUpdateMembershipType];

export const MemberUpdateMembershipType = {
  basic: "basic",
  standard: "standard",
  premium: "premium",
} as const;

export type MemberUpdateStatus =
  (typeof MemberUpdateStatus)[keyof typeof MemberUpdateStatus];

export const MemberUpdateStatus = {
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export interface MemberUpdate {
  name?: string;
  email?: string;
  phone?: string;
  membershipType?: MemberUpdateMembershipType;
  membershipExpiry?: string;
  status?: MemberUpdateStatus;
  /** @nullable */
  workoutPlanId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Attendance {
  id: number;
  memberId: number;
  memberName: string;
  checkInAt: string;
  /** @nullable */
  checkOutAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface AttendanceInput {
  memberId: number;
  /** @nullable */
  notes?: string | null;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  paid: "paid",
  pending: "pending",
  overdue: "overdue",
} as const;

export interface Payment {
  id: number;
  memberId: number;
  memberName: string;
  amount: number;
  status: PaymentStatus;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export type PaymentInputStatus =
  (typeof PaymentInputStatus)[keyof typeof PaymentInputStatus];

export const PaymentInputStatus = {
  paid: "paid",
  pending: "pending",
  overdue: "overdue",
} as const;

export interface PaymentInput {
  memberId: number;
  amount: number;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  status?: PaymentInputStatus;
}

export type PaymentUpdateStatus =
  (typeof PaymentUpdateStatus)[keyof typeof PaymentUpdateStatus];

export const PaymentUpdateStatus = {
  paid: "paid",
  pending: "pending",
  overdue: "overdue",
} as const;

export interface PaymentUpdate {
  amount?: number;
  status?: PaymentUpdateStatus;
  /** @nullable */
  description?: string | null;
  dueDate?: string;
  /** @nullable */
  paidAt?: string | null;
}

export type WorkoutPlanLevel =
  (typeof WorkoutPlanLevel)[keyof typeof WorkoutPlanLevel];

export const WorkoutPlanLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  /** @nullable */
  day?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface WorkoutPlan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  level: WorkoutPlanLevel;
  durationWeeks: number;
  exercises: Exercise[];
  createdAt: string;
}

export type WorkoutPlanInputLevel =
  (typeof WorkoutPlanInputLevel)[keyof typeof WorkoutPlanInputLevel];

export const WorkoutPlanInputLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface WorkoutPlanInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  level: WorkoutPlanInputLevel;
  durationWeeks: number;
  exercises?: Exercise[];
}

export type WorkoutPlanUpdateLevel =
  (typeof WorkoutPlanUpdateLevel)[keyof typeof WorkoutPlanUpdateLevel];

export const WorkoutPlanUpdateLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface WorkoutPlanUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  level?: WorkoutPlanUpdateLevel;
  durationWeeks?: number;
  exercises?: Exercise[];
}

export interface Branding {
  id: number;
  gymName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  primaryColor: string;
  secondaryColor: string;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  themeName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface BrandingUpdate {
  gymName?: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  primaryColor?: string;
  secondaryColor?: string;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  themeName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  expiredMembers: number;
  todayAttendance: number;
  monthlyRevenue: number;
  pendingPayments: number;
  totalWorkoutPlans: number;
  newMembersThisMonth: number;
}

export interface RevenueDataPoint {
  month: string;
  revenue: number;
}

export interface AttendanceDataPoint {
  date: string;
  count: number;
}

export type ListMembersParams = {
  status?: ListMembersStatus;
  search?: string;
};

export type ListMembersStatus =
  (typeof ListMembersStatus)[keyof typeof ListMembersStatus];

export const ListMembersStatus = {
  active: "active",
  expired: "expired",
  suspended: "suspended",
} as const;

export type ListAttendanceParams = {
  memberId?: number;
  date?: string;
};

export type ListPaymentsParams = {
  memberId?: number;
  status?: ListPaymentsStatus;
};

export type ListPaymentsStatus =
  (typeof ListPaymentsStatus)[keyof typeof ListPaymentsStatus];

export const ListPaymentsStatus = {
  paid: "paid",
  pending: "pending",
  overdue: "overdue",
} as const;