# zip-repl/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GymFlow Gym Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, suspended]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Attendance ─────────────────────────────────────────────
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: memberId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Record member check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: memberId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, pending, overdue]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─── Workout Plans ──────────────────────────────────────────
...
components:
  schemas:
...
    Member:
      type: object
      required: [id, name, email, phone, membershipType, membershipExpiry, status, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        membershipType:
          type: string
          enum: [basic, standard, premium]
        membershipExpiry:
          type: string
          format: date
        status:
          type: string
          enum: [active, expired, suspended]
        joinedAt:
          type: string
          format: date-time
        workoutPlanId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    MemberInput:
      type: object
      required: [name, email, phone, membershipType, membershipExpiry]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        membershipType:
          type: string
          enum: [basic, standard, premium]
        membershipExpiry:
          type: string
          format: date
        workoutPlanId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        membershipType:
          type: string
          enum: [basic, standard, premium]
        membershipExpiry:
          type: string
          format: date
        status:
          type: string
          enum: [active, expired, suspended]
        workoutPlanId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    Attendance:
      type: object
      required: [id, memberId, memberName, checkInAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
    AttendanceInput:
      type: object
      required: [memberId]
      properties:
        memberId:
          type: integer
        notes:
          type: ["string", "null"]
    Payment:
      type: object
      required: [id, memberId, memberName, amount, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue]
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [memberId, amount, dueDate]
      properties:
        memberId:
          type: integer
        amount:
          type: number
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [paid, pending, overdue]
    PaymentUpdate:
      type: object
      properties:
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue]
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date
        paidAt:
          type: ["string", "null"]
...